Split value.h
This file was recompiled the most in the entire project.
