+++ src/c_api/connection.cpp
@@ -1,9 +1,14 @@
 #include "binder/bound_statement_result.h"
 #include "c_api/kuzu.h"
 #include "common/exception/exception.h"
-#include "common/types/value.h"
 #include "main/kuzu.h"
 #include "planner/logical_plan/logical_plan.h"
+
+namespace kuzu {
+namespace common {
+class Value;
+}
+} // namespace kuzu
 
 using namespace kuzu::common;
 using namespace kuzu::main;
