+++ src/expression_evaluator/case_evaluator.cpp
@@ -1,5 +1,11 @@
 #include "expression_evaluator/case_evaluator.h"
 
+#include "common/types/date_t.h"
+#include "common/types/interval_t.h"
+#include "common/types/ku_string.h"
+#include "common/types/timestamp_t.h"
+#include "common/types/types.h"
+
 using namespace kuzu::common;
 using namespace kuzu::processor;
 using namespace kuzu::storage;
