+++ src/include/common/types/value/rel.h
@@ -1,0 +1,72 @@
+#include <cstdint>
+#include <memory>
+#include <string>
+#include <utility>
+#include <vector>
+
+#include "common/api.h"
+#include "common/types/internal_id_t.h"
+
+namespace kuzu {
+namespace common {
+
+class Value;
+
+/**
+ * @brief RelVal represents a rel in the graph and stores the relID, src/dst nodes and properties of
+ * that rel.
+ */
+class RelVal {
+public:
+    /**
+     * @return all properties of the RelVal.
+     * @note this function copies all the properties into a vector, which is not efficient. use
+     * `getPropertyName` and `getPropertyVal` instead if possible.
+     */
+    KUZU_API static std::vector<std::pair<std::string, std::unique_ptr<Value>>> getProperties(
+        const Value* val);
+    /**
+     * @return number of properties of the RelVal.
+     */
+    KUZU_API static uint64_t getNumProperties(const Value* val);
+    /**
+     * @return the name of the property at the given index.
+     */
+    KUZU_API static std::string getPropertyName(const Value* val, uint64_t index);
+    /**
+     * @return the value of the property at the given index.
+     */
+    KUZU_API static Value* getPropertyVal(const Value* val, uint64_t index);
+    /**
+     * @return the src nodeID value of the RelVal in Value.
+     */
+    KUZU_API static Value* getSrcNodeIDVal(const Value* val);
+    /**
+     * @return the dst nodeID value of the RelVal in Value.
+     */
+    KUZU_API static Value* getDstNodeIDVal(const Value* val);
+    /**
+     * @return the src nodeID value of the RelVal as nodeID struct.
+     */
+    KUZU_API static nodeID_t getSrcNodeID(const Value* val);
+    /**
+     * @return the dst nodeID value of the RelVal as nodeID struct.
+     */
+    KUZU_API static nodeID_t getDstNodeID(const Value* val);
+    /**
+     * @return the name of the RelVal.
+     */
+    KUZU_API static std::string getLabelName(const Value* val);
+    /**
+     * @return the value of the RelVal in string format.
+     */
+    KUZU_API static std::string toString(const Value* val);
+
+private:
+    static void throwIfNotRel(const Value* val);
+    // 4 offset for id, label, src, dst.
+    static constexpr uint64_t OFFSET = 4;
+};
+
+} // namespace common
+} // namespace kuzu
