+++ src/include/storage/storage_structure/column.h
@@ -2,7 +2,7 @@
 
 #include "catalog/catalog.h"
 #include "common/exception/storage.h"
-#include "common/types/value.h"
+#include "common/types/value/value.h"
 #include "storage/storage_structure/disk_overflow_file.h"
 #include "storage/storage_structure/storage_structure.h"
