+++ src/planner/operator/logical_dummy_scan.cpp
@@ -1,5 +1,7 @@
 #include "planner/logical_plan/scan/logical_dummy_scan.h"
 
+#include "common/constants.h"
+
 using namespace kuzu::common;
 
 namespace kuzu {
