+++ src/processor/result/flat_tuple.cpp
@@ -2,6 +2,7 @@
 
 #include "common/exception/runtime.h"
 #include "common/string_utils.h"
+#include "common/types/value/value.h"
 #include "utf8proc.h"
 #include "utf8proc_wrapper.h"
