+++ src/storage/copier/column_chunk.cpp
@@ -1,6 +1,7 @@
 #include "storage/copier/column_chunk.h"
 
 #include "arrow/array.h"
+#include "common/types/value/nested.h"
 #include "storage/copier/string_column_chunk.h"
 #include "storage/copier/struct_column_chunk.h"
 #include "storage/copier/table_copy_utils.h"
