+++ src/storage/copier/struct_column_chunk.cpp
@@ -4,6 +4,7 @@
 #include "common/exception/not_implemented.h"
 #include "common/exception/parser.h"
 #include "common/string_utils.h"
+#include "common/types/value/nested.h"
 #include "storage/copier/string_column_chunk.h"
 #include "storage/copier/table_copy_utils.h"
 #include "storage/copier/var_list_column_chunk.h"
