+++ src/storage/storage_structure/in_mem_file.cpp
@@ -5,6 +5,7 @@
 #include "common/exception/copy.h"
 #include "common/exception/storage.h"
 #include "common/type_utils.h"
+#include "common/types/value/nested.h"
 
 using namespace kuzu::common;
