+++ tools/java_api/src/jni/kuzu_java.cpp
@@ -6,7 +6,10 @@
 #include "com_kuzudb_KuzuNative.h"
 #include "common/exception/exception.h"
 #include "common/types/types.h"
-#include "common/types/value.h"
+#include "common/types/value/value.h"
+#include "common/types/value/nested.h"
+#include "common/types/value/node.h"
+#include "common/types/value/rel.h"
 #include "json.hpp"
 #include "main/kuzu.h"
 #include "main/query_summary.h"
